Keep trace contracts independently runnable
Constraint: Lead requested scoped trace verification without running the full HTTP or bare test suite.
Rejected: Leaving trace coverage embedded in forum-traces.test.ts | tests/http/traces/ would remain unavailable as a scoped target.
Confidence: high
Scope-risk: narrow
Directive: Add future trace HTTP contracts under tests/http/traces/.
Tested: bun test tests/http/traces/; bun test tests/http/forum-traces.test.ts; tsc --noEmit.
Not-tested: full bare bun test intentionally skipped because it pulls agents/ worktrees.
